requirements.txt
Django==5.0.4
django-filter==24.2
openpyxl==3.1.2
djangorestframework==3.15.1
django-cors-headers==4.3.1
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
